requirements.txt
azure-functions
snowflake-connector-python
mysql-connector